fixed 2 tests that failed in the UserTest class

+++ src/test/java/data/UserTest.java
@@ -110,7 +110,7 @@ public class UserTest {
     @Test
     public void getElectricityDailyConsumptionSuccess(){
         userOne.setElectricityDailyConsumption(70000);
-        Assert.assertEquals(70000, userOne.getElectricityDailyConsumption());
+        Assert.assertEquals(70000, (int) userOne.getElectricityDailyConsumption());
     }
 
     @Test
@@ -121,7 +121,7 @@ public class UserTest {
     @Test
     public void setElectricityDailyConsumptionSuccsess() {
         userOne.setElectricityDailyConsumption(50000);
-        Assert.assertEquals(50000, userOne.getElectricityDailyConsumption());
+        Assert.assertEquals(50000,(int) userOne.getElectricityDailyConsumption());
     }
 
     @Test
